requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
python-telegram-bot==13.15
pytz==2024.2
requests==2.32.3
six==1.17.0
tinydb==4.8.2
tornado==6.1
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
